+++ src/views/LoanDetails/LoanDetails.js
@@ -475,7 +475,7 @@ class LoanDetails extends React.Component {
                     </Button>
                   </IfPermission>
                 }
-                <IfPermission perm="ui-users.loans.edit">
+                <IfPermission perm="ui-users.loans.change-due-date">
                   <Button
                     data-test-change-due-date-button
                     disabled={