+++ src/components/Loans/OpenLoans/components/ActionsDropdown/ActionsDropdown.js
@@ -88,7 +88,7 @@ class ActionsDropdown extends React.Component {
             </Button>
           }
         </IfPermission>
-        <IfPermission perm="ui-users.loans.edit">
+        <IfPermission perm="ui-users.loans.change-due-date">
           { itemStatusName !== itemStatuses.DECLARED_LOST &&
             itemStatusName !== itemStatuses.CLAIMED_RETURNED &&
             itemStatusName !== itemStatuses.AGED_TO_LOST &&