+++ src/components/Accounts/Actions/FeeFineActions.js
@@ -79,7 +79,7 @@ class Actions extends React.Component {
     transfers: {
       type: 'okapi',
       records: 'transfers',
-      path: 'transfers?limit=100',
+      path: 'transfers?limit=2000',
     },
     curUserServicePoint: {
       type: 'okapi',