+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Show an error message, not a spinner, when a loan is missing. Refs UIU-1045.
 * Remove "Users: User loan edit" permission.
 * Add "Users: User loans change due date" permission.
+* Show correct source of fee/fine in payment 'action' for 'Charge & pay now'. Refs UIU-1981.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.0...v5.0.1)