+++ CHANGELOG.md
@@ -33,6 +33,8 @@
 * Create manual patron block templates in settings. Refs UIU-1909.
 * Show an error message, not a spinner, when a loan is missing. Refs UIU-1045.
 * Populate expiration date based on pre-defined offset for patron group. Refs UIU-1907.
+* Remove "Users: User loan edit" permission.
+* Add "Users: User loans change due date" permission.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.0...v5.0.1)