+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Populate expiration date based on pre-defined offset for patron group. Refs UIU-1907.
 * Remove "Users: User loan edit" permission.
 * Add "Users: User loans change due date" permission.
+* Show correct source of fee/fine in payment 'action' for 'Charge & pay now'. Refs UIU-1981.
 * Change `limit=100` to `limit=2000` across the board; patrons have more than 100 things. All. The. Time. Refs UIU-1987.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)