+++ CHANGELOG.md
@@ -32,7 +32,11 @@
 * Validate 'Expiration date offset (days)' of patron groups. Refs UIU-1951.
 * Create manual patron block templates in settings. Refs UIU-1909.
 * Show an error message, not a spinner, when a loan is missing. Refs UIU-1045.
-* Change `limit=100` to `limit=1000` across the board; patrons have more than 100 things. All. The. Time. Refs UIU-1987.
+* Remove "Users: User loan edit" permission.
+* Add "Users: User loans change due date" permission.
+* Show correct source of fee/fine in payment 'action' for 'Charge & pay now'. Refs UIU-1981.
+* Change `limit=100` to `limit=2000` across the board; patrons have more than 100 things. All. The. Time. Refs UIU-1987.
+* Fix the selection of service points from the `Add service points` popup to display in the user record. Fixes UIU-1912.
 * Retrieve up to 10k tags instead of 10. Refs UIU-2003
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)